Split out OS specific code into separate modules
